+++ keyhunt.c
@@ -10,6 +10,7 @@ email: alberto.bsd@gmail.com
 #include <string.h>
 #include <unistd.h>
 #include <pthread.h>
+#include <math.h>
 #include <time.h>
 #include "keccak/keccak-tiny.h"
 #include "base58/libbase58.h"
@@ -53,10 +54,14 @@ void Point_Addition(struct Point *P, struct Point *Q, struct Point *R);
 void Scalar_Multiplication(struct Point P, struct Point *R, mpz_t m);
 void Point_Negation(struct Point A, struct Point *S);
 int searchbinary(char *BUFFER,char *data,int length,int _N);
-void quicksort(char *arr, int low, int high);
-int partition (char *arr, int low, int high);
-void swap(char *a,char *b);
 
+void _sort(char *arr,int N);
+void _insertionsort(char *arr, int n);
+void _introsort(char *arr,int depthLimit, int n);
+void swap(char *a,char *b);
+int partition(char *arr, int n);
+void heapsort(char  *arr, int n);
+void heapify(char *arr, int n, int i);
 
 void *thread_process(void *vargp);
 void *thread_process_range(void *vargp);
@@ -485,10 +490,12 @@ int main(int argc, char **argv)	{
 	if(FLAGALREADYSORTED)	{
 	  printf("File mark already sorted, skipping sort proccess\n");
 		printf("%i values were loaded\n",N);
+		_insertionsort(DATABUFFER,N);
 	}
 	else	{
 		printf("sorting data\n");
-		quicksort(DATABUFFER,0,N-1);
+		_sort(DATABUFFER,N);
+		_insertionsort(DATABUFFER,N);
 		printf("%i values were loaded and sorted\n",N);
 	}
 
@@ -719,6 +726,7 @@ void init_doublingG(struct Point *P)	{
 		i++;
 	}
 }
+
 char *pubkeytopubaddress_eth(char *pkey,int length)	{
 		char *temp,*pubaddress = calloc(MAXLENGTHADDRESS,1);
 		char *digest = malloc(32);
@@ -1229,27 +1237,147 @@ void swap(char *a,char *b)  {
   memcpy(b,t,MAXLENGTHADDRESS);
 }
 
-int partition (char *arr, int low, int high)  {
-    char *pivot = arr + (high*MAXLENGTHADDRESS);    // pivot
-		//printf("Pivot : %s\n",pivot);
-    int j,i = (low - 1);  // Index of smaller element
-    for (j = low; j < high; j++)  {
-        // If current element is smaller than the pivot
-        if (memcmp(arr + (j*MAXLENGTHADDRESS),pivot,MAXLENGTHADDRESS) < 0)  {
-            i++;    // increment index of smaller element
-            swap(arr + (i*MAXLENGTHADDRESS), arr + (j*MAXLENGTHADDRESS));
-        }
-    }
-    swap(arr + ((i+1)*MAXLENGTHADDRESS), arr + (high*MAXLENGTHADDRESS));
-    return (i + 1);
+void _sort(char *arr,int n)  {
+  int depthLimit = ((int) ceil(log(n))) * 2;
+  _introsort(arr,depthLimit,n);
 }
 
-void quicksort(char *arr, int low, int high)  {
-  int pi;
-  if (low < high)  {
-			//printf("quicksort from %i to %i\n",low,high);
-      pi = partition(arr, low, high);
-      quicksort(arr, low, pi - 1);
-      quicksort(arr, pi + 1, high);
+void _introsort(char *arr,int depthLimit, int n) {
+  int p;
+  if(n > 1)  {
+    if(n <= 16) {
+      _insertionsort(arr,n);
+    }
+    else  {
+      if(depthLimit == 0) {
+        heapsort(arr,n);
+      }
+      else  {
+        p = partition(arr,n);
+        if(p >= 2) {
+          _introsort(arr , depthLimit-1 , p);
+        }
+        if((n - (p + 1)) >= 2 ) {
+          _introsort(arr + ((p+1) *MAXLENGTHADDRESS) , depthLimit-1 , n - (p + 1));
+        }
+      }
+    }
+  }
+}
+
+void _insertionsort(char *arr, int n) {
+	int j,i;
+  char *arrj,*temp;
+  char key[MAXLENGTHADDRESS];
+  for(i = 1; i < n ; i++ ) {
+    j= i-1;
+    memcpy(key,arr + (i*MAXLENGTHADDRESS),MAXLENGTHADDRESS);
+    arrj = arr + (j*MAXLENGTHADDRESS);
+    while(j >= 0 && memcmp(arrj,key,MAXLENGTHADDRESS) > 0) {
+      memcpy(arr + ((j+1)*MAXLENGTHADDRESS),arrj,MAXLENGTHADDRESS);
+      j--;
+      arrj = arr + (j*MAXLENGTHADDRESS);
+    }
+    memcpy(arr + ((j+1)*MAXLENGTHADDRESS),key,MAXLENGTHADDRESS);
+  }
+}
+
+int partition(char *arr, int n)  {
+  char pivot[MAXLENGTHADDRESS];
+  int j,i,t, r = (int) n/2,jaux = -1,iaux = -1, iflag, jflag;
+  char *a,*b,*hextemp,*hextemp_pivot;
+  i = - 1;
+  memcpy(pivot,arr + (r*MAXLENGTHADDRESS),MAXLENGTHADDRESS);
+  i = 0;
+  j = n-1;
+  do {
+    iflag = 1;
+    jflag = 1;
+    t = memcmp(arr + (i*MAXLENGTHADDRESS),pivot,MAXLENGTHADDRESS);
+    iflag = (t <= 0);
+    while(i < j && iflag) {
+      i++;
+      t = memcmp(arr + (i*MAXLENGTHADDRESS),pivot,MAXLENGTHADDRESS);
+      iflag = (t <= 0);
+    }
+    t = memcmp(arr + (j*MAXLENGTHADDRESS),pivot,MAXLENGTHADDRESS);
+    jflag = (t > 0);
+    while(i < j && jflag) {
+      j--;
+      t = memcmp(arr + (j*MAXLENGTHADDRESS),pivot,MAXLENGTHADDRESS);
+      jflag = (t > 0);
+    }
+    if(i < j) {
+      if(i == r )  {
+        r = j;
+      }
+      else  {
+        if(j == r )  {
+          r = i;
+        }
+      }
+
+      swap(arr + (i*MAXLENGTHADDRESS),arr + (j*MAXLENGTHADDRESS) );
+      jaux = j;
+      iaux = i;
+      j--;
+      i++;
+    }
+
+  } while(j > i );
+  if(jaux != -1 && iaux != -1)  {
+    if(iflag || jflag)  {
+      if(iflag) {
+        if(r != j)
+          swap(arr + (r*MAXLENGTHADDRESS),arr + ((j )*MAXLENGTHADDRESS) );
+        jaux = j;
+      }
+      if(jflag) {
+        if(r != j-1)
+          swap(arr + (r*MAXLENGTHADDRESS),arr + ((j-1 )*MAXLENGTHADDRESS) );
+        jaux = j-1;
+      }
+    }
+    else{
+      if(r != j)
+        swap(arr + (r*MAXLENGTHADDRESS),arr + ((j )*MAXLENGTHADDRESS) );
+      jaux = j;
+    }
+  }
+  else  {
+    if(iflag && jflag)  {
+      jaux = r;
+    }
+    else  {
+      if(iflag ) {
+        swap(arr + (r*MAXLENGTHADDRESS),arr + ((j)*MAXLENGTHADDRESS) );
+        jaux = j;
+      }
+    }
+  }
+  return jaux;
+}
+
+void heapify(char *arr, int n, int i) {
+    int largest = i;
+    int l = 2 * i + 1;
+    int r = 2 * i + 2;
+    if (l < n && memcmp(arr +(l*MAXLENGTHADDRESS),arr +(largest * MAXLENGTHADDRESS),MAXLENGTHADDRESS) > 0)
+        largest = l;
+    if (r < n && memcmp(arr +(r*MAXLENGTHADDRESS),arr +(largest *MAXLENGTHADDRESS),MAXLENGTHADDRESS) > 0)
+        largest = r;
+    if (largest != i) {
+        swap(arr +(i*MAXLENGTHADDRESS), arr +(largest*MAXLENGTHADDRESS));
+        heapify(arr, n, largest);
+    }
+}
+
+void heapsort(char  *arr, int n)  {
+  int i;
+  for ( i = n / 2 - 1; i >= 0; i--)
+    heapify(arr, n, i);
+  for ( i = n - 1; i > 0; i--) {
+    swap(arr , arr +(i*MAXLENGTHADDRESS));
+    heapify(arr, i, 0);
   }
 }